+++ tests/integration_tests/features.py
@@ -11,9 +11,7 @@ import subprocess
 from dataclasses import dataclass
 from typing import List, Sequence
 
-
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+from torchtitan.tools.logging import logger
 
 
 @dataclass
@@ -31,7 +29,7 @@ class OverrideDefinitions:
         return self.test_descr
 
 
-def build_test_list() -> List[OverrideDefinitions]:
+def build_features_test_list() -> List[OverrideDefinitions]:
     """
     key is the config file name and value is a list of OverrideDefinitions
     that is used to generate variations of integration tests based on the
@@ -552,103 +550,3 @@ def build_test_list() -> List[OverrideDefinitions]:
         ),
     ]
     return integration_tests_flavors
-
-
-def _run_cmd(cmd):
-    return subprocess.run([cmd], text=True, shell=True)
-
-
-def run_single_test(test_flavor: OverrideDefinitions, full_path: str, output_dir: str):
-    # run_test supports sequence of tests.
-    test_name = test_flavor.test_name
-    dump_folder_arg = f"--job.dump_folder {output_dir}/{test_name}"
-
-    all_ranks = ",".join(map(str, range(test_flavor.ngpu)))
-
-    for idx, override_arg in enumerate(test_flavor.override_args):
-        cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_train.sh"
-        # dump compile trace for debugging purpose
-        cmd = f'TORCH_TRACE="{output_dir}/{test_name}/compile_trace" ' + cmd
-        if test_name == "fsdp2_memory_estimation":
-            cmd = (
-                f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} "
-                "./scripts/estimate/run_memory_estimation.sh"
-            )
-        cmd += " " + dump_folder_arg
-        if override_arg:
-            cmd += " " + " ".join(override_arg)
-        logger.info(
-            f"=====Integration test, flavor : {test_flavor.test_descr}, command : {cmd}====="
-        )
-
-        # save checkpoint (idx == 0) and load it for generation (idx == 1)
-        if test_name == "test_generate" and idx == 1:
-            cmd = (
-                f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} "
-                f"CHECKPOINT_DIR={output_dir}/{test_name}/checkpoint/step-10 "
-                "PROMPT='What is the meaning of life?' "
-                f"./scripts/generate/run_llama_generate.sh --out > {output_dir}/{test_name}/generated_output.json"
-            )
-
-        result = _run_cmd(cmd)
-        logger.info(result.stdout)
-        if result.returncode != 0:
-            raise Exception(
-                f"Integration test failed, flavor : {test_flavor.test_descr}, command : {cmd}"
-            )
-
-
-def run_tests(args):
-    """Run all integration tests to test the core features of TorchTitan"""
-
-    test_list = build_test_list()
-    for test_flavor in test_list:
-        # Filter by test_name if specified
-        if args.test_name != "all" and test_flavor.test_name != args.test_name:
-            continue
-
-        # Check if config file exists
-        assert args.config_path.endswith(
-            ".toml"
-        ), "Base config path must end with .toml"
-        assert os.path.exists(
-            args.config_path
-        ), f"Base config path {args.config_path} does not exist"
-
-        # Check if we have enough GPUs
-        if args.ngpu < test_flavor.ngpu:
-            logger.info(
-                f"Skipping test {test_flavor.test_name} that requires {test_flavor.ngpu} gpus,"
-                f" because --ngpu arg is {args.ngpu}"
-            )
-        else:
-            run_single_test(test_flavor, args.config_path, args.output_dir)
-
-
-def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument("output_dir", help="Directory to store test outputs")
-    parser.add_argument(
-        "--config_path",
-        default="./tests/integration_tests/base_config.toml",
-        help="Base config path for integration tests. This is the config that will be used as a base for all tests.",
-    )
-    parser.add_argument(
-        "--test_name",
-        default="all",
-        help="Specific test name to run (e.g., 'tp_only', 'full_checkpoint'). Use 'all' to run all tests (default: all)",
-    )
-    parser.add_argument(
-        "--ngpu", default=8, type=int, help="Maximum number of GPUs to use"
-    )
-    args = parser.parse_args()
-
-    if not os.path.exists(args.output_dir):
-        os.makedirs(args.output_dir)
-    if os.listdir(args.output_dir):
-        raise RuntimeError("Please provide an empty output directory.")
-    run_tests(args)
-
-
-if __name__ == "__main__":
-    main()
