+++ tests/integration_tests/run_tests.py
@@ -1,0 +1,134 @@
+import argparse
+import logging
+import os
+import subprocess
+from dataclasses import dataclass
+from typing import List, Sequence
+
+from tests.integration_tests.features import build_features_test_list
+from tests.integration_tests.ft import build_ft_test_list
+from tests.integration_tests.h100 import build_h100_tests_list
+from tests.integration_tests.models import build_model_tests_list
+from torchtitan.tools.logging import logger
+
+from .features import OverrideDefinitions
+
+
+_TEST_SUITES_FUNCTION = {
+    "features": build_features_test_list,
+    "ft": build_ft_test_list,
+    "models": build_model_tests_list,
+    "h100": build_h100_tests_list,
+}
+
+
+def _run_cmd(cmd):
+    return subprocess.run([cmd], text=True, shell=True)
+
+
+def run_single_test(test_flavor: OverrideDefinitions, full_path: str, output_dir: str):
+    # run_test supports sequence of tests.
+    test_name = test_flavor.test_name
+    dump_folder_arg = f"--job.dump_folder {output_dir}/{test_name}"
+
+    all_ranks = ",".join(map(str, range(test_flavor.ngpu)))
+
+    for idx, override_arg in enumerate(test_flavor.override_args):
+        cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_train.sh"
+        # dump compile trace for debugging purpose
+        cmd = f'TORCH_TRACE="{output_dir}/{test_name}/compile_trace" ' + cmd
+        if test_name == "fsdp2_memory_estimation":
+            cmd = (
+                f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} "
+                "./scripts/estimate/run_memory_estimation.sh"
+            )
+
+        cmd += " " + dump_folder_arg
+        if override_arg:
+            cmd += " " + " ".join(override_arg)
+        logger.info(
+            f"=====Integration test, flavor : {test_flavor.test_descr}, command : {cmd}====="
+        )
+
+        # save checkpoint (idx == 0) and load it for generation (idx == 1)
+        if test_name == "test_generate" and idx == 1:
+            cmd = (
+                f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} "
+                f"CHECKPOINT_DIR={output_dir}/{test_name}/checkpoint/step-10 "
+                "PROMPT='What is the meaning of life?' "
+                f"./scripts/generate/run_llama_generate.sh --out > {output_dir}/{test_name}/generated_output.json"
+            )
+
+        result = _run_cmd(cmd)
+        logger.info(result.stdout)
+        if result.returncode != 0:
+            raise Exception(
+                f"Integration test failed, flavor : {test_flavor.test_descr}, command : {cmd}"
+            )
+
+
+def run_tests(args, test_list: List[OverrideDefinitions]):
+    """Run all integration tests to test the core features of TorchTitan"""
+
+    for test_flavor in test_list:
+        # Filter by test_name if specified
+        if args.test_name != "all" and test_flavor.test_name != args.test_name:
+            continue
+
+        # Check if config file exists
+        assert args.config_path.endswith(
+            ".toml"
+        ), "Base config path must end with .toml"
+        assert os.path.exists(
+            args.config_path
+        ), f"Base config path {args.config_path} does not exist"
+
+        # Check if we have enough GPUs
+        if args.ngpu < test_flavor.ngpu:
+            logger.info(
+                f"Skipping test {test_flavor.test_name} that requires {test_flavor.ngpu} gpus,"
+                f" because --ngpu arg is {args.ngpu}"
+            )
+        else:
+            run_single_test(test_flavor, args.config_path, args.output_dir)
+
+
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("output_dir", help="Directory to store test outputs")
+    parser.add_argument(
+        "--test_suite",
+        default="",
+        choices=["features", "ft", "models", "h100"],
+        help="Which test suite to run. If not specified, torchtitan composibility tests will be run",
+    )
+    parser.add_argument(
+        "--config_path",
+        default="./tests/integration_tests/base_config.toml",
+        help="Base config path for integration tests. This is the config that will be used as a base for all tests.",
+    )
+    parser.add_argument(
+        "--test_name",
+        default="all",
+        help="Specific test name to run (e.g., 'tp_only', 'full_checkpoint'). Use 'all' to run all tests (default: all)",
+    )
+    parser.add_argument(
+        "--ngpu", default=8, type=int, help="Maximum number of GPUs to use"
+    )
+    args = parser.parse_args()
+
+    if not os.path.exists(args.output_dir):
+        os.makedirs(args.output_dir)
+    if os.listdir(args.output_dir):
+        raise RuntimeError("Please provide an empty output directory.")
+
+    assert (
+        args.test_suite in _TEST_SUITES_FUNCTION
+    ), f"Unknown test suite {args.test_suite}"
+
+    test_list = _TEST_SUITES_FUNCTION[args.test_suite]()
+    run_tests(args, test_list)
+
+
+if __name__ == "__main__":
+    main()
