+++ torchtitan/experiments/flux/tests/integration_tests.py
@@ -10,150 +10,113 @@ import os
 import subprocess
 
 from tests.integration_tests.features import OverrideDefinitions
+from tests.integration_tests.run_tests import run_tests
 
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+from torchtitan.tools.logging import logger
 
 
-def build_test_list():
+def build_flux_test_list():
     """
     key is the config file name and value is a list of OverrideDefinitions
     that is used to generate variations of integration tests based on the
     same root config file.
     """
-    integration_tests_flavors = []
-    integration_tests_flavors.extend(
-        [
-            # basic tests
-            OverrideDefinitions(
+    integration_tests_flavors = [
+        # basic tests
+        OverrideDefinitions(
+            [
                 [
-                    [
-                        "--profiling.enable_profiling",
-                        "--metrics.enable_tensorboard",
-                    ],
+                    "--model.name flux",
+                    "--training.test_mode",
+                    "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/",
+                    "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/",
+                    "--model.tokenizer_path tests/assets/tokenizer",
+                    "--profiling.enable_profiling",
+                    "--metrics.enable_tensorboard",
                 ],
-                "default",
-                "default",
-            ),
-            # Checkpointing tests.
-            OverrideDefinitions(
+            ],
+            "default",
+            "default",
+        ),
+        # Checkpointing tests.
+        OverrideDefinitions(
+            [
                 [
-                    [
-                        "--checkpoint.enable_checkpoint",
-                    ],
-                    [
-                        "--checkpoint.enable_checkpoint",
-                        "--training.steps 20",
-                    ],
+                    "--model.name flux",
+                    "--training.test_mode",
+                    "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/",
+                    "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/",
+                    "--model.tokenizer_path tests/assets/tokenizer",
+                    "--checkpoint.enable_checkpoint",
                 ],
-                "Checkpoint Integration Test - Save Load Full Checkpoint",
-                "full_checkpoint",
-            ),
-            OverrideDefinitions(
                 [
-                    [
-                        "--checkpoint.enable_checkpoint",
-                        "--checkpoint.last_save_model_only",
-                    ],
+                    "--model.name flux",
+                    "--training.test_mode",
+                    "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/",
+                    "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/",
+                    "--model.tokenizer_path tests/assets/tokenizer",
+                    "--checkpoint.enable_checkpoint",
+                    "--training.steps 20",
                 ],
-                "Checkpoint Integration Test - Save Model Only fp32",
-                "last_save_model_only_fp32",
-            ),
-            # Parallelism tests.
-            OverrideDefinitions(
+            ],
+            "Checkpoint Integration Test - Save Load Full Checkpoint",
+            "full_checkpoint",
+        ),
+        OverrideDefinitions(
+            [
                 [
-                    [
-                        "--parallelism.data_parallel_shard_degree=4",
-                        "--parallelism.data_parallel_replicate_degree=1",
-                    ]
+                    "--model.name flux",
+                    "--training.test_mode",
+                    "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/",
+                    "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/",
+                    "--model.tokenizer_path tests/assets/tokenizer",
+                    "--checkpoint.enable_checkpoint",
+                    "--checkpoint.last_save_model_only",
                 ],
-                "FSDP",
-                "fsdp",
-                ngpu=4,
-            ),
-            OverrideDefinitions(
+            ],
+            "Checkpoint Integration Test - Save Model Only fp32",
+            "last_save_model_only_fp32",
+        ),
+        # Parallelism tests.
+        OverrideDefinitions(
+            [
                 [
-                    [
-                        "--parallelism.data_parallel_shard_degree=2",
-                        "--parallelism.data_parallel_replicate_degree=2",
-                    ]
-                ],
-                "HSDP",
-                "hsdp",
-                ngpu=4,
-            ),
-        ]
-    )
+                    "--model.name flux",
+                    "--training.test_mode",
+                    "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/",
+                    "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/",
+                    "--model.tokenizer_path tests/assets/tokenizer",
+                    "--parallelism.data_parallel_shard_degree=4",
+                    "--parallelism.data_parallel_replicate_degree=1",
+                ]
+            ],
+            "FSDP",
+            "fsdp",
+            ngpu=4,
+        ),
+        OverrideDefinitions(
+            [
+                [
+                    "--model.name flux",
+                    "--training.test_mode",
+                    "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/",
+                    "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/",
+                    "--model.tokenizer_path tests/assets/tokenizer",
+                    "--parallelism.data_parallel_shard_degree=2",
+                    "--parallelism.data_parallel_replicate_degree=2",
+                ]
+            ],
+            "HSDP",
+            "hsdp",
+            ngpu=4,
+        ),
+    ]
     return integration_tests_flavors
 
 
-def _run_cmd(cmd):
-    return subprocess.run([cmd], text=True, shell=True)
-
-
-def run_single_test(test_flavor: OverrideDefinitions, full_path: str, output_dir: str):
-    # run_test supports sequence of tests.
-    test_name = test_flavor.test_name
-    dump_folder_arg = f"--job.dump_folder {output_dir}/{test_name}"
-
-    # Random init encoder for offline testing
-    random_init_encoder_arg = "--training.test_mode"
-    clip_encoder_version_arg = "--encoder.clip_encoder torchtitan/experiments/flux/tests/assets/clip-vit-large-patch14/"
-    t5_encoder_version_arg = (
-        "--encoder.t5_encoder torchtitan/experiments/flux/tests/assets/t5-v1_1-xxl/"
-    )
-    hf_assets_path_arg = "--model.hf_assets_path tests/assets/tokenizer"
-
-    all_ranks = ",".join(map(str, range(test_flavor.ngpu)))
-
-    for idx, override_arg in enumerate(test_flavor.override_args):
-        cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./torchtitan/experiments/flux/run_train.sh"
-        # dump compile trace for debugging purpose
-        cmd = f'TORCH_TRACE="{output_dir}/{test_name}/compile_trace" ' + cmd
-        cmd += " " + dump_folder_arg
-        cmd += " " + random_init_encoder_arg
-        cmd += " " + clip_encoder_version_arg
-        cmd += " " + t5_encoder_version_arg
-        cmd += " " + hf_assets_path_arg
-        if override_arg:
-            cmd += " " + " ".join(override_arg)
-        logger.info(
-            f"=====Flux Integration test, flavor : {test_flavor.test_descr}, command : {cmd}====="
-        )
-
-        result = _run_cmd(cmd)
-        logger.info(result.stdout)
-        if result.returncode != 0:
-            raise Exception(
-                f"Flux Integration test failed, flavor : {test_flavor.test_descr}, command : {cmd}"
-            )
-
-
-def run_tests(args):
-    # build integration tests list
-    test_list = build_test_list()
-
-    for test_flavor in test_list:
-        # Filter by test_name if specified
-        if args.test_name != "all" and test_flavor.test_name != args.test_name:
-            continue
-
-        # Check if config file exists
-        assert args.config_path.endswith(
-            ".toml"
-        ), "Base config path must end with .toml"
-        assert os.path.exists(
-            args.config_path
-        ), f"Base config path {args.config_path} does not exist"
-
-        # Check if we have enough GPUs
-        if args.ngpu < test_flavor.ngpu:
-            logger.info(
-                f"Skipping test {test_flavor.test_name} that requires {test_flavor.ngpu} gpus,"
-                f" because --ngpu arg is {args.ngpu}"
-            )
-        else:
-            run_single_test(test_flavor, args.config_path, args.output_dir)
+_TEST_SUITES_FUNCTION = {
+    "flux": build_flux_test_list,
+}
 
 
 def main():
@@ -176,7 +139,9 @@ def main():
         os.makedirs(args.output_dir)
     if os.listdir(args.output_dir):
         raise RuntimeError("Please provide an empty output directory.")
-    run_tests(args)
+
+    test_list = _TEST_SUITES_FUNCTION["flux"]()
+    run_tests(args, test_list)
 
 
 if __name__ == "__main__":
